pyproject.toml
[tool.poetry]
name = "kpi_api"
packages = [
    { include = "kpi_api", from = "src" }
]
version = "0.1.0"
description = "API for Grafana"
authors = ["qberal <quentin.beral@insa-rouen.fr>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "^0.7.0"
anyio = "^4.8.0"
caldav = "^1.4.0"
certifi = "^2024.12.14"
charset-normalizer = "^3.4.1"
click = "^8.1.8"
config = "^0.5.1"
fastapi = "^0.115.7"
greenlet = "^3.1.1"
h11 = "^0.14.0"
icalendar = "^6.1.1"
idna = "^3.10"
lxml = "^5.3.0"
numpy = "^2.2.2"
opencv-python = "^4.11.0.86"
playwright = "^1.49.1"
pydantic = "^2.10.6"
pydantic-core = "^2.27.2"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
pytz = "^2024.2"
recurring-ical-events = "^3.4.1"
requests = "^2.32.3"
six = "^1.17.0"
sniffio = "^1.3.1"
starlette = "^0.45.3"
typing-extensions = "^4.12.2"
tzdata = "^2025.1"
urllib3 = "^2.3.0"
uvicorn = "^0.34.0"
vobject = "^0.9.9"
x-wr-timezone = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
